requirements.txt
anyio==3.3.0
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
contextlib2==21.6.0
docopt==0.6.2
ffmpeg-python==0.2.0
future==0.18.2
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==3.2
internetarchive==2.1.0
jsonpatch==1.32
jsonpointer==2.1
Pillow==8.3.2
pkg-resources==0.0.0
PyYAML==5.4.1
requests==2.26.0
rfc3986==1.5.0
schema==0.7.4
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
sqlitedict==1.7.0
telegraph==1.4.1
tqdm==4.62.2
typing-extensions==3.10.0.0
urllib3==1.26.6
